refactor: move assets into src/ folder
