+++ src/assets/animations/index.ts
