+++ src/components/FileInput.tsx
@@ -4,7 +4,7 @@ import {Text, StyleSheet, Image, TouchableOpacity, Alert} from 'react-native';
 import {Buffer} from 'buffer';
 
 import * as RNFS from 'react-native-fs';
-import {fileIcon} from '../../assets/icons';
+import {fileIcon} from '../assets/icons';
 
 export interface PickedFile {
   name: string;
