+++ src/index.js
@@ -4,6 +4,5 @@
 
 import {AppRegistry} from 'react-native';
 import App from './App';
-import {name as appName} from '../app.json';
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent('fax-app', () => App);
